Add shared bar history utilities
